Added a 4-byte int version number to the start of the output file. Since the old file (version 1) had the image width in the first 4 bytes, the backend can detect version 1 files by checking for a version # greater than, say, 319, since it's unlikely that at video would have been recorded at a smaller width than that.
Added 4-byte int cursor coords to the start of each frame. These are only
valid if a system cursor is being used (which won't appear in the rendered
frames). If a GL cursor is being used (which *will* appear in the rendered
frames) then the coords will be (-1, -1). This allows the backend to add a
cursor image to those frames from which it is absent. In Minecraft system
cursors are used in all of the GUIs while a GL cursor is used in the main
3D view.

Set the version to 2.